requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
flake8==3.9.2
mccabe==0.6.1
Pillow==9.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.6.0
python-decouple==3.1
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.2
whitenoise==4.1.2
gunicorn==20.0.4
dj-database-url==0.5.0